+++ src/pretalx/orga/forms/cfp.py
@@ -160,6 +160,21 @@ class AnswerOptionForm(ReadOnlyFlag, I18nModelForm):
 
 
 class SubmissionTypeForm(ReadOnlyFlag, I18nModelForm):
+    def __init__(self, *args, event=None, **kwargs):
+        self.event = event
+        super().__init__(*args, **kwargs)
+
+    def clean_name(self):
+        name = self.cleaned_data["name"]
+        qs = self.event.submission_types.all()
+        if self.instance and self.instance.pk:
+            qs = qs.exclude(pk=self.instance.pk)
+        if any(str(s.name) == str(name) for s in qs):
+            raise forms.ValidationError(
+                _("You already have a submission type by this name!")
+            )
+        return name
+
     def save(self, *args, **kwargs):
         instance = super().save(*args, **kwargs)
         if instance.pk and "duration" in self.changed_data:
@@ -174,9 +189,19 @@ class SubmissionTypeForm(ReadOnlyFlag, I18nModelForm):
 
 
 class TrackForm(ReadOnlyFlag, I18nModelForm):
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args, event=None, **kwargs):
+        self.event = event
         super().__init__(*args, **kwargs)
         self.fields["color"].widget.attrs["class"] = "colorpickerfield"
+
+    def clean_name(self):
+        name = self.cleaned_data["name"]
+        qs = self.event.tracks.all()
+        if self.instance and self.instance.pk:
+            qs = qs.exclude(pk=self.instance.pk)
+        if any(str(s.name) == str(name) for s in qs):
+            raise forms.ValidationError(_("You already have a track by this name!"))
+        return name
 
     class Meta:
         model = Track